usb: dwc3: add support for role reversal
with this patch dwc3 can switch roles without
any known issues.

Note that it's not necessary to reset the IP,
all we have to do is cache and restore a few
registers while switching roles. So far, all
known versions of the IP share the physical
RAM0 space, so all current versions will need
this workaround. If Synopsys decides to change
that behavior, a revision check should be added
for the register save/restore bit.

[ balbi@ti.com :
	- rewritten on top of recent kernel;
	- made it generic so non-OMAP could use
	- added comments to code
	- improved commit log a bit ]

Signed-off-by: George Cherian <george.cherian@ti.com>
Signed-off-by: Felipe Balbi <balbi@ti.com>