arm: omap: hwmod: allow for registration of class-less hwmods
Before this patch, HWMOD requires the existence
of a struct omap_hwmod_class very early. It just
so happens that we're moving that data to DeviceTree
in a follow-up commit so we need to allow the
registration of HWMODs with a valid omap_hwmod_class
pointer.

When the device is built from DT, we will allocate and
populate struct omap_hwmod_class based on DT data.

Signed-off-by: Felipe Balbi <balbi@ti.com>