phy: introduce bus power control methods
sometimes we might want to manually control
a bus power rail. For example, USB provides
a VBUS line which, in some implementations,
need to be manually controlled by the OS.

One such case is Texas Instruments' AM437x
where we need to cope with the possibility
of VBUS contention in SW because HW won't
be able to handle it for us.

Tested-by: Bin Liu <b-liu@ti.com>
Signed-off-by: Felipe Balbi <balbi@ti.com>