usb: dwc3: implement VBUS contention protection
Before entering host mode, we need to check if VBUS
is already valid. If it is, then we can't enter host
mode otherwise we will have two entities driving VBUS
in a single USB wire.

Likewise, we need to make sure that after enabling
VBUS power, VBUS actually turned on and raised above
VBUS Valid threshold, otherwise we have a VBUS Error
condition which we should also cope with.

Tested-by: Bin Liu <b-liu@ti.com>
Signed-off-by: Felipe Balbi <balbi@ti.com>